Adding field 'size' to the alloccell_t structure of the packed type representation. This prepares the structure for the implementation of the PUSH, POP and TRIM opcodes, and makes this implementation more similar to the other two representations (full stackcell and split stackcell).
git-svn-id: svn+ssh://saulius.grazulis.lt/home/saulius/svn-repositories/compilers/trunk@3027 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ sc/interpreters/packed-type/bcalloc.c
@@ -54,7 +54,7 @@ static struct {
 };
 #endif
 
-void *bcalloc( short element_size, ssize_t length, ssize_t nref )
+void *bcalloc( ssize_t element_size, ssize_t length, ssize_t nref )
 {
     alloccell_t *ptr = NULL;
     ssize_t size = (length < 0 ? 1 : length) * element_size +
